requirements.txt
anyio==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dadata==21.10.1
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
pyTelegramBotAPI==4.15.4
python-dotenv==1.0.1
python-telegram-bot==20.8
requests==2.31.0
sniffio==1.3.0
telebot==0.0.5
typing_extensions==4.9.0
urllib3==2.2.0
